Binning optimization (#11)
* updating directory name from tools to bin_opt

* debugging technical issues using run 2 datacard

* bin opt legacy files

* some python specific format changes

* law task for binning optimization

* adding the worker command

* debugging

* resetting the ref_bkgs

* hardsetting environment

* using subprocess call function

* changing the binning hot fix

* removing negative binning print outs

* modifying for Run 3 2022 preEE

* configurable yaml file

* updating configurable files

* removing txt files for zero/negative integral processes

* adding function to adjust bins on histogram to be equidistant

* histogram reading bug fix

* adding a sample stack plot script

* adding hint and fixes for adding a binning suggestion json file

* bug fix to call the previous datacards within a channel that have been optimized

* plotting script update

* removing law configuration from this PR

* removing rename processes script

* Re-adding the option to use multiple acquisition functions

+++ bin_opt/rebinAndRunLimitsWorker.py
@@ -4,44 +4,54 @@ import shutil
 import time
 import uuid
 
+import yaml
+input_binning_opt_config = os.path.join(os.environ["ANALYSIS_PATH"], "StatInference", "bin_opt", "bin_optimization.yaml")
+with open(input_binning_opt_config, "r") as f:
+    input_binning_opt_config_dict = yaml.safe_load(f)
+
 from rebinAndRunLimits import GetLimits
 
-import argparse
+for channel in input_binning_opt_config_dict["input"]["channels"]:
+    if os.path.isfile(os.path.join(input_binning_opt_config_dict["output"]["directory"],
+                                   f'{channel}_{str(input_binning_opt_config_dict["input"].get("era", ""))}.json')):
+                                   print( f"Results for channel {channel} already exist.")
+    else:
+        worker_dir = os.path.join(input_binning_opt_config_dict["output"]["directory"],
+        f'{channel}_{str(input_binning_opt_config_dict["input"].get("era", ""))}',
+                                            "workers",
+                                            uuid.uuid4().hex)
 
-parser = argparse.ArgumentParser(description='Rebin histogram and run expected limits.')
-parser.add_argument('--output', required=True, type=str, help="output directory")
-parser.add_argument('--verbose', required=False, type=int, default=1, help="verbosity")
-args = parser.parse_args()
+        os.makedirs(worker_dir, exist_ok=True)
+        task_file = os.path.join(worker_dir, 'task.txt')
+        result_file = os.path.join(worker_dir, 'result.txt')
+        result_file_tmp = os.path.join(worker_dir, '.result.txt')
 
-worker_dir = os.path.join(args.output, uuid.uuid4().hex)
-os.mkdir(worker_dir)
-task_file = os.path.join(worker_dir, 'task.txt')
-result_file = os.path.join(worker_dir, 'result.txt')
-result_file_tmp = os.path.join(worker_dir, '.result.txt')
+        verbose = input_binning_opt_config_dict.get("verbose", 1)
 
-if args.verbose > 0:
-    print('Worker dir: {}'.format(worker_dir))
+        if verbose > 0:
+            print('Worker dir: {}'.format(worker_dir))
 
-while True:
-    time.sleep(1)
-    if os.path.isfile(task_file) and not os.path.isfile(result_file):
-        try:
-            with open(task_file, 'r') as f:
-                params = json.load(f)
-        except IOError:
-            continue
-        bin_edges = params['bin_edges']
-        if args.verbose > 0:
-            print('Bin edges: [ {} ]'.format(', '.join([ str(b) for b in bin_edges ])))
-        limit = GetLimits(params['input_datacard'], worker_dir, bin_edges, params['poi'], verbose=0,
-                          other_datacards=params['other_datacards'])
-        if args.verbose > 0:
-            print('Expected 95% CL limit = {}'.format(limit))
-        result = {
-            'input_datacard': params['input_datacard'],
-            'bin_edges': bin_edges,
-            'exp_limit': limit
-        }
-        with open(result_file_tmp, 'w') as f:
-            json.dump(result, f)
-        shutil.move(result_file_tmp, result_file)
+        while True:
+            time.sleep(1)
+            if os.path.isfile(task_file) and not os.path.isfile(result_file):
+                try:
+                    with open(task_file, 'r') as f:
+                        params = json.load(f)
+                except IOError:
+                    continue
+                bin_edges = params['bin_edges']
+                if verbose > 0:
+                    print('Bin edges: [ {} ]'.format(', '.join([ str(b) for b in bin_edges ])))
+                limit = GetLimits(params['input_datacard'], worker_dir, bin_edges, params['poi'], verbose=1,
+                                    other_datacards=params['other_datacards'])
+
+                if verbose > 0:
+                    print('Expected 95% CL limit = {}'.format(limit))
+                result = {
+                    'input_datacard': params['input_datacard'],
+                    'bin_edges': bin_edges,
+                    'exp_limit': limit
+                }
+                with open(result_file_tmp, 'w') as f:
+                    json.dump(result, f)
+                shutil.move(result_file_tmp, result_file)
